[dagit] Make the “Latest Run” tag on the Job page auto-refresh
